task-3/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
sentence-transformers==2.2.2
torch==2.2.1
langchain==0.2.12
langgraph==0.2.19
langchain-google-genai==1.0.6
chromadb==0.5.5
joblib==1.4.2
scikit-learn==1.5.2
xgboost==2.1.1
pandas==2.2.3
numpy==2.1.2
redis==5.1.0
boto3==1.35.51
matplotlib==3.9.2
seaborn==0.13.2
